feat(types): add src/types/ with navigation, pet, and component types, update MyDonationScreen imports
