+++ src/components/customText/CustomText.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View, Text } from 'react-native';
 import styles from './style';
-import { CustomTextProps } from '../../types/types';
+import { CustomTextProps } from 'types';
 const CustomText: React.FC<CustomTextProps> = ({ title, style }) => {
   return (
     <View>
