+++ src/components/imgWithText/ImgWithText.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {View, Text, Image} from 'react-native';
 import styles from './style';
-import {ImgWithTextProps} from '../../types/types';
+import {ImgWithTextProps} from 'types';
 const ImgWithText: React.FC<ImgWithTextProps> = ({imageSource, label}) => {
   return (
     <View style={styles.container}>
