+++ src/hooks/usePetDetails.ts
@@ -3,7 +3,7 @@ import {useAppDispatch} from '../hooks/hooks';
 import auth from '@react-native-firebase/auth';
 import {useNavigation} from '@react-navigation/native';
 import {StackNavigationProp} from '@react-navigation/stack';
-import {RootStackParamList, Pet, ProfileData} from '../types/types';
+import {RootStackParamList, Pet, ProfileData} from 'types';
 import {requestAdoption} from '../redux/slices/petSlice';
 import firestore from '@react-native-firebase/firestore';
 import Toast from 'react-native-toast-message';
