+++ src/navigation/AppNavigator.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { createDrawerNavigator } from '@react-navigation/drawer';
 import CustomDrawer from '../navigation/DrawerNavigation';
 import { Screen } from '../constants/screen'; 
-import { DrawerParamList } from '../types/types';
+import { DrawerParamList } from 'types';
 const Drawer = createDrawerNavigator<DrawerParamList>();
 const AppNavigator = () => {
   return (
