+++ src/navigation/AppStack.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
-import {AppStackParamList} from '../types/types';
+import {AppStackParamList} from 'types';
 import {AppScreens} from '../constants/screen';
 const Stack = createNativeStackNavigator<AppStackParamList>();
 const AppStack = () => {
