+++ src/screens/adoptNowScreen/style.ts
@@ -5,7 +5,10 @@ const styles = StyleSheet.create({
     flex: 1,
     backgroundColor: COLOR.white,
     paddingHorizontal: 17,
+
+
   },
+  
   cardContainer: {
     marginTop: 20,
     marginBottom: 30,
