+++ src/screens/favouriteScreen/hook.ts
@@ -1,0 +1,63 @@
+import { useEffect } from 'react';
+import Toast from 'react-native-toast-message';
+import useFavorites from '../../hooks/useFavourite'; 
+import { usePetDetails } from '../../hooks/usePetDetails';
+
+const useFavouriteScreen = () => {
+  const {
+    favorites,
+    selectedPet,
+    isModalVisible,
+    handlePetClick,
+    toggleFavorite,
+    closeModal,
+    loading,
+    error,
+    noFavoritesMessage,
+  } = useFavorites();
+
+  const { profileData, handleAdoptNow } = usePetDetails(selectedPet);
+
+  // Show error toast when error occurs
+  useEffect(() => {
+    if (error) {
+      Toast.show({
+        type: 'error',
+        text1: 'Favorites Error',
+        text2: error,
+      });
+    }
+  }, [error]);
+
+  // Show info toast when there are no favorites and not loading
+  useEffect(() => {
+    if (noFavoritesMessage && !loading) {
+      Toast.show({
+        type: 'info',
+        text1: 'Favorites',
+        text2: noFavoritesMessage,
+      });
+    }
+  }, [noFavoritesMessage, loading]);
+
+  // Handler for toggling favorite
+  const handleToggleFavorite = (pet) => {
+    toggleFavorite(pet);
+  };
+
+  return {
+    favorites,
+    selectedPet,
+    isModalVisible,
+    handlePetClick,
+    handleToggleFavorite,
+    closeModal,
+    loading,
+    error,
+    noFavoritesMessage,
+    profileData,
+    handleAdoptNow,
+  };
+};
+
+export default useFavouriteScreen;
